+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-# Changelog
-
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [v0.1.3] - 2024-03-16
+
+### Added
+- Initial owner set to "not_yet_owned" in Smart Contracts (SC).
+- Pricing & buying method in SC.
+- Buying method for a specific pixel in frontend.
+- Clickable pixel on canvas.
 
 ## [v0.1.2] - 2024-03-15
 
@@ -37,3 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - First version with small canvas, fetching their colors in the SC datastore.
+
+---
+
+Make sure to replace `[date]` with the actual date you intend to release version v0.1.3.